requirements.txt
celery==5.2.7
channels==3.0.5
channels-redis==3.4.0
coreschema==0.0.4
Django==4.0.5
django-allauth==0.50.0
django-ckeditor==6.4.2
django-filter==21.1
django-money==3.0.0
django-rest-friendship==0.2.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
friend==0.2.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py-moneyed==2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
